+++ chat-widget/src/components/webchatcontainerstateful/interfaces/IAdaptiveCardStyles.ts
@@ -1,0 +1,4 @@
+export interface IAdaptiveCardStyles {
+    background?: string;
+    color?: string;
+}