+++ chat-widget/src/contexts/common/LiveChatWidgetActionType.ts
@@ -206,4 +206,10 @@ export enum LiveChatWidgetActionType {
         any: Set height and width of Widget
     */
     SET_WIDGET_SIZE,
+
+    /*
+        Parameters:
+        any: Set widget instance id
+    */
+    SET_WIDGET_INSTANCE_ID,
 }